🚚 reorganized Home Component into a single directory
