+++ apps/ui/src/app/components/home/imports.ts
@@ -1,0 +1,7 @@
+export type { IconModel } from '../../imports/models';
+
+export { IconStub } from '../../imports/constants';
+
+export { RouterService } from '../../imports/services';
+
+export { RouteConstants } from '../../app.routes';
