+++ apps/ui/src/app/components/home/index.ts
@@ -1,0 +1,8 @@
+export type { HomeAssetModel, HomeAssetsModel } from './model/home.model';
+
+export { HomeAssets, HomeAssetsStub } from './constant/home.constants';
+
+export { HomeContainerComponent } from './component/home-container/home-container.component';
+export { HomeComponent } from './component/home-container/home/home.component';
+
+export { HomeService } from './service/home.service';
