+++ apps/ui/src/app/components/index.ts
@@ -1,2 +1,1 @@
-export { HomeContainerComponent } from './home/home-container/home-container.component';
-export { HomeComponent } from './home/home-container/home/home.component';
+export { HomeContainerComponent, HomeComponent } from './home';
