--- apps/ui/src/app/constants/index.ts
@@ -1,1 +1,0 @@
-export { HomeAssets, HomeAssetsStub } from './home/home.constants';
