--- apps/ui/src/app/models/index.ts
@@ -1,4 +1,0 @@
-export type {
-  HomeAssetModel,
-  HomeAssetsModel
-} from './home/home.model';
