--- apps/ui/src/app/services/index.ts
@@ -1,1 +1,0 @@
-export { HomeService } from './home/home.service';
